+++ pytest/common_lib/shared/mpc_cluster.py
@@ -1,218 +1,60 @@
 import base64
-import base58
-import os
-import sys
-import json
-import time
-import yaml
 import hashlib
+import json
 import pathlib
-import subprocess
+import sys
+import time
 from concurrent.futures import ThreadPoolExecutor
-from ruamel.yaml import YAML
+
+import requests
+
 from common_lib import constants
 from common_lib import signature
+from common_lib.constants import TGAS, TIMEOUT
 from common_lib.contract_state import ContractState, ProtocolState, SignatureScheme
+from common_lib.shared.mpc_node import MpcNode
+from common_lib.shared.near_node import NearNode
 from common_lib.signature import generate_sign_args
 
 sys.path.append(str(pathlib.Path(__file__).resolve().parents[1]))
 
-from cluster import LocalNode
 from typing import List, Optional
 
-from cluster import start_cluster
-from utils import MetricsTracker
-
-from transaction import create_create_account_action, create_payment_action, \
-                        create_full_access_key_action, sign_deploy_contract_tx, \
-                        sign_transaction, serialize_transaction, sign_function_call_tx
-
-from key import Key
-
-import requests
-from .constants import NEAR_BASE, MPC_BINARY_PATH, TGAS, TIMEOUT
+from transaction import sign_deploy_contract_tx
 
 
-# Some boilerplate to make pyyaml ignore unknown fields
-def ignore_unknown(loader, tag_suffix, node):
-    return None
-
-
-class SafeLoaderIgnoreUnknown(yaml.SafeLoader):
-    pass
-
-
-SafeLoaderIgnoreUnknown.add_multi_constructor('!', ignore_unknown)
-
-
-def assert_txn_success(res):
-    assert 'result' in res, json.dumps(res, indent=1)
-    assert 'status' in res['result'], json.dumps(res['result'], indent=1)
-    assert 'SuccessValue' in res['result']['status'], json.dumps(
-        res['result']['status'])
-
-
-class NearNode:
-
-    def __init__(self, near_node: LocalNode):
-        self.near_node = near_node
-
-    def signer_key(self):
-        return self.near_node.signer_key
-
-    def account_id(self):
-        return self.signer_key().account_id
-
-    def last_block_hash(self):
-        return self.near_node.get_latest_block().hash_bytes
-
-    def send_tx(self, txn):
-        return self.near_node.send_tx(txn)
-
-    def get_tx(self, tx_hash):
-        return self.near_node.get_tx(tx_hash, self.account_id())
-
-    def send_txn_and_check_success(self, txn, timeout=20):
-        res = self.near_node.send_tx_and_wait(txn, timeout)
-        assert_txn_success(res)
-        return res
-
-    def get_nonce(self):
-        return self.near_node.get_nonce_for_pk(
-            self.near_node.signer_key.account_id, self.near_node.signer_key.pk)
-
-    def sign_tx(self,
-                target_contract,
-                function_name,
-                args,
-                nonce_offset=1,
-                gas=150 * TGAS,
-                deposit=0):
-        last_block_hash = self.last_block_hash()
-        nonce = self.get_nonce() + nonce_offset
-        encoded_args = args if type(args) == bytes else json.dumps(
-            args).encode('utf-8')
-        tx = sign_function_call_tx(self.signer_key(), target_contract,
-                                   function_name, encoded_args, gas, deposit,
-                                   nonce, last_block_hash)
-        return tx
-
-
-class MpcNode(NearNode):
-
-    class NodeStatus:
-        # not a participant, neither a candidate
-        IDLE = 1
-        # a participant in the current epoch and also in the next epoch
-        PARTICIPANT = 2
-        # a participant in the current epoch but not in the next epoch
-        OLD_PARTICIPANT = 3
-        # a participant in the next epoch, but not the current epoch
-        NEW_PARTICIPANT = 4
-
-    def __init__(self, near_node: LocalNode, url, sign_pk):
-        super().__init__(near_node)
-        self.account_id = near_node.signer_key.account_id
-        self.url = url
-        self.sign_pk = sign_pk
-        self.status = MpcNode.NodeStatus.IDLE
-        self.participant_id = None
-        self.home_dir = self.near_node.node_dir
-        self.is_running = False
-        self.metrics = MetricsTracker(near_node)
-
-    def change_contract_id(self, new_contract_id: str):
-        yaml = YAML()
-        yaml.preserve_quotes = True  # optional: keeps any quotes if present in original file
-
-        path = pathlib.Path(self.home_dir) / 'config.yaml'
-        with path.open('r') as f:
-            config = yaml.load(f)
-
-        old_contract_id = config['indexer']['mpc_contract_id']
+def verify_txs(results, verification_callback, verbose=False):
+    max_tgas_used = 0
+    total_tgas = 0
+    total_receipts = 0
+    num_txs = 0
+    for res in results:
+        num_txs += 1
+        gas_tx, n_rcpts_tx = extract_tx_costs(res)
+        max_tgas_used = max(max_tgas_used, gas_tx) / TGAS
+        total_tgas += gas_tx / TGAS
+        total_receipts += n_rcpts_tx
+        verification_callback(res)
+    if verbose:
         print(
-            f"changing contract_id from {old_contract_id} to {new_contract_id} for node {self.account_id}"
+            f"number of txs: {num_txs}\n max gas used (Tgas):{max_tgas_used}\n average receipts: {total_receipts / num_txs}\n average gas used (Tgas): {total_tgas / num_txs}\n"
         )
-        config['indexer']['mpc_contract_id'] = new_contract_id
 
-        with path.open('w') as f:
-            yaml.dump(config, f)
 
-    def print(self):
-        if not self.is_running:
-            return f"⛔\033[90m{self.account_id}\033[0m"
-        if self.status == MpcNode.NodeStatus.IDLE:
-            return f"⚫\033[90m{self.account_id}\033[0m"
-        if self.status == MpcNode.NodeStatus.PARTICIPANT:
-            return f"🟢\033[92m{self.account_id}\033[0m"
-        if self.status == MpcNode.NodeStatus.OLD_PARTICIPANT:
-            return f"🟢\033[92m{self.account_id}\033[0m"
-        if self.status == MpcNode.NodeStatus.NEW_PARTICIPANT:
-            return f"🟡\033[93m{self.account_id}\033[0m"
-        else:
-            return f"❓\033[0m{self.account_id}\033[0m"
+def extract_tx_costs(res):
+    """
+    returns `total_gas_used`, `num_receipts`
+    """
+    # Extract the gas burnt at transaction level
+    total_gas_used = res["result"]["transaction_outcome"]["outcome"][
+        "gas_burnt"]
 
-    def set_secret_store_key(self, secret_store_key):
-        self.secret_store_key = secret_store_key
-
-    def reset_mpc_data(self):
-        assert not self.is_running
-        patterns = [
-            'CURRENT',
-            'IDENTITY'
-            'LOCK',
-            'LOG',
-            'MANIFEST-.*',
-            'OPTIONS-.*',
-            '*.log',
-            '*.sst',
-        ]
-        for pattern in patterns:
-            for file_path in pathlib.Path(self.home_dir).glob(pattern):
-                file_path.unlink()
-
-    def run(self):
-        assert not self.is_running
-        self.is_running = True
-
-        p2p_private_key = open(pathlib.Path(self.home_dir) / 'p2p_key').read()
-        near_secret_key = json.loads(
-            open(pathlib.Path(self.home_dir) /
-                 'validator_key.json').read())['secret_key']
-        extra_env = {
-            'RUST_LOG': 'INFO',  # mpc-node produces too much output on DEBUG
-            'MPC_SECRET_STORE_KEY': self.secret_store_key,
-            'MPC_P2P_PRIVATE_KEY': p2p_private_key,
-            'MPC_ACCOUNT_SK': near_secret_key,
-        }
-        cmd = (MPC_BINARY_PATH, 'start', '--home-dir', self.home_dir)
-        self.near_node.run_cmd(cmd=cmd, extra_env=extra_env)
-
-    def kill(self, gentle=True):
-        self.near_node.kill(gentle=gentle)
-        self.is_running = False
-
-    def wait_for_connection_count(self, awaited_count):
-        started = time.time()
-        while True:
-            assert time.time(
-            ) - started < TIMEOUT, "Waiting for connection count"
-            try:
-                conns = self.metrics.get_metric_all_values(
-                    "mpc_network_live_connections")
-                print("mpc_network_live_connections", conns)
-                connection_count = int(sum([kv[1] for kv in conns]))
-                if connection_count == awaited_count:
-                    break
-            except requests.exceptions.ConnectionError:
-                pass
-            time.sleep(1)
-
-    def reserve_key_event_attempt(self, epoch_id, domain_id, attempt_id):
-        file_path = pathlib.Path(self.home_dir)
-        file_path = file_path / "temporary_keys" / f"started_{epoch_id}_{domain_id}_{attempt_id}"
-        file_path.parent.mkdir(parents=True, exist_ok=True)
-        file_path.touch()
+    # Add the gas burnt for each receipt
+    num_receipts = 0
+    for receipt in res["result"]["receipts_outcome"]:
+        total_gas_used += receipt["outcome"]["gas_burnt"]
+        num_receipts += 1
+    return total_gas_used, num_receipts
 
 
 class MpcCluster:
@@ -240,8 +82,8 @@ class MpcCluster:
     def get_voters(self):
         voters = [
             node for node in self.mpc_nodes if node.is_running and (
-                node.status == MpcNode.NodeStatus.OLD_PARTICIPANT
-                or node.status == MpcNode.NodeStatus.PARTICIPANT)
+                    node.status == MpcNode.NodeStatus.OLD_PARTICIPANT
+                    or node.status == MpcNode.NodeStatus.PARTICIPANT)
         ]
         print("Voters:", " | ".join([node.print() for node in voters]))
         return voters
@@ -355,7 +197,7 @@ class MpcCluster:
             'threshold': threshold,
             'participants': {
                 'next_id':
-                self.next_participant_id,
+                    self.next_participant_id,
                 'participants': [[
                     node.account_id, node.participant_id, {
                         'sign_pk': node.sign_pk,
@@ -640,192 +482,3 @@ class MpcCluster:
                           'remove_timed_out_requests',
                           {'max_num_to_remove': max_num_to_remove})
         return node.send_txn_and_check_success(tx)
-
-
-def extract_tx_costs(res):
-    """
-    returns `total_gas_used`, `num_receipts`
-    """
-    # Extract the gas burnt at transaction level
-    total_gas_used = res["result"]["transaction_outcome"]["outcome"][
-        "gas_burnt"]
-
-    # Add the gas burnt for each receipt
-    num_receipts = 0
-    for receipt in res["result"]["receipts_outcome"]:
-        total_gas_used += receipt["outcome"]["gas_burnt"]
-        num_receipts += 1
-    return total_gas_used, num_receipts
-
-
-def verify_txs(results, verification_callback, verbose=False):
-    max_tgas_used = 0
-    total_tgas = 0
-    total_receipts = 0
-    num_txs = 0
-    for res in results:
-        num_txs += 1
-        gas_tx, n_rcpts_tx = extract_tx_costs(res)
-        max_tgas_used = max(max_tgas_used, gas_tx) / TGAS
-        total_tgas += gas_tx / TGAS
-        total_receipts += n_rcpts_tx
-        verification_callback(res)
-    if verbose:
-        print(
-            f"number of txs: {num_txs}\n max gas used (Tgas):{max_tgas_used}\n average receipts: {total_receipts / num_txs}\n average gas used (Tgas): {total_tgas / num_txs}\n"
-        )
-
-
-# Output is deserializable into the rust type near_crypto::SecretKey
-def serialize_key(key: Key):
-    full_key = bytes(key.decoded_sk())
-    return 'ed25519:' + base58.b58encode(full_key).decode('ascii')
-
-
-def sign_create_account_with_multiple_access_keys_tx(creator_key,
-                                                     new_account_id, keys,
-                                                     nonce, block_hash):
-    create_account_action = create_create_account_action()
-    payment_action = create_payment_action(100 * NEAR_BASE)
-    access_key_actions = [
-        create_full_access_key_action(key.decoded_pk()) for key in keys
-    ]
-    actions = [create_account_action, payment_action] + access_key_actions
-    signed_tx = sign_transaction(new_account_id, nonce, actions, block_hash,
-                                 creator_key.account_id,
-                                 creator_key.decoded_pk(),
-                                 creator_key.decoded_sk())
-    return serialize_transaction(signed_tx)
-
-
-def start_cluster_with_mpc(num_validators,
-                           num_mpc_nodes,
-                           num_respond_aks,
-                           contract,
-                           presignatures_to_buffer=None,
-                           start_mpc_nodes=True):
-    rpc_polling_config = {
-        "rpc": {
-            "polling_config": {
-                "polling_timeout": {
-                    "secs": 20,
-                    "nanos": 0
-                },
-                "polling_interval": {
-                    "secs": 1,
-                    "nanos": 0
-                },
-            }
-        }
-    }
-
-    # Start a near network with extra observer nodes; we will use their
-    # config.json, genesis.json, etc. to configure the mpc nodes' indexers
-
-    nodes = start_cluster(
-        num_validators, num_mpc_nodes, 1, None,
-        [["epoch_length", 1000], ["block_producer_kickout_threshold", 80]], {
-            0: rpc_polling_config,
-            1: rpc_polling_config
-        })
-
-    mpc_node_indices = range(num_validators, num_validators + num_mpc_nodes)
-    for i in mpc_node_indices:
-        nodes[i].kill(gentle=True)
-        nodes[i].reset_data()
-
-    # Generate the mpc configs
-    dot_near = pathlib.Path.home() / '.near'
-    cmd = (MPC_BINARY_PATH, 'generate-test-configs', '--output-dir', dot_near,
-           '--participants', ','.join(f'test{i + num_validators}'
-                                      for i in range(num_mpc_nodes)),
-           '--threshold', str(num_mpc_nodes))
-    if presignatures_to_buffer:
-        cmd = cmd + (
-            '--desired-presignatures-to-buffer',
-            str(presignatures_to_buffer),
-        )
-    subprocess.run(cmd)
-
-    # Get the participant set from the mpc configs.
-    candidates = []
-    with open(pathlib.Path(dot_near / 'participants.json')) as file:
-        participants_config = yaml.load(file, Loader=SafeLoaderIgnoreUnknown)
-    for i, p in enumerate(participants_config['participants']):
-        near_account = p['near_account_id']
-        assert near_account == f"test{i + num_validators}", \
-            f"This test only works with account IDs 'testX' where X is the node index; expected 'test{i + num_validators}', got {near_account}"
-        my_pk = p['p2p_public_key']
-        my_addr = p['address']
-        my_port = p['port']
-
-        candidates.append({
-            "account_id":
-            near_account,
-            "cipher_pk": [
-                0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-                0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
-            ],
-            "sign_pk":
-            my_pk,
-            "url":
-            f"http://{my_addr}:{my_port}",
-        })
-    for i in mpc_node_indices:
-        # Move the generated mpc configs
-        mpc_config_dir = dot_near / str(i - num_validators)
-        for fname in os.listdir(mpc_config_dir):
-            subprocess.run(('mv', os.path.join(mpc_config_dir,
-                                               fname), nodes[i].node_dir))
-
-    mpc_nodes = [
-        MpcNode(nodes[i], candidates[i - num_validators]["url"],
-                candidates[i - num_validators]["sign_pk"])
-        for i in mpc_node_indices
-    ]
-    cluster = MpcCluster(near_nodes=[NearNode(node) for node in nodes])
-
-    last_block_hash = cluster.contract_node.last_block_hash()
-    # Set up the node's home directories
-    for mpc_node in mpc_nodes:
-        # Indexer config must explicitly specify tracked shard
-        fname = os.path.join(mpc_node.near_node.node_dir, 'config.json')
-        with open(fname) as fd:
-            config_json = json.load(fd)
-        config_json['tracked_shards'] = [0]
-        with open(fname, 'w') as fd:
-            json.dump(config_json, fd, indent=2)
-        print(f"Wrote {fname} as config for node {mpc_node.account_id}")
-
-        # Create respond.yaml with credentials for sending responses
-        if num_respond_aks > 0:
-            account_id = f"respond.{mpc_node.account_id}"
-            access_keys = [
-                Key.from_seed_testonly(account_id, seed=f"{s}")
-                for s in range(0, num_respond_aks)
-            ]
-            tx = sign_create_account_with_multiple_access_keys_tx(
-                mpc_node.signer_key(), account_id, access_keys, 1,
-                last_block_hash)
-            cluster.contract_node.send_txn_and_check_success(tx)
-            respond_cfg = {
-                'account_id': account_id,
-                'access_keys': list(map(serialize_key, access_keys)),
-            }
-            fname = os.path.join(mpc_node.near_node.node_dir, 'respond.yaml')
-            with open(fname, "w") as file:
-                yaml.dump(respond_cfg, file, default_flow_style=False)
-
-    # Deploy the mpc contract
-    cluster.deploy_contract(contract)
-
-    # Name mpc nodes A, B, C, ...
-    for i, mpc_node in enumerate(mpc_nodes):
-        mpc_node.set_secret_store_key(str(chr(ord('A') + i) * 32))
-
-    # Start the mpc nodes
-    if start_mpc_nodes:
-        for mpc_node in mpc_nodes:
-            mpc_node.run()
-
-    return cluster, mpc_nodes
