+++ pytest/common_lib/shared/yaml_safeloader.py
@@ -1,0 +1,13 @@
+import yaml
+
+
+# Some boilerplate to make pyyaml ignore unknown fields
+def ignore_unknown(loader, tag_suffix, node):
+    return None
+
+
+class SafeLoaderIgnoreUnknown(yaml.SafeLoader):
+    pass
+
+
+SafeLoaderIgnoreUnknown.add_multi_constructor('!', ignore_unknown)
